Preserve DICE
This program was written to make life easier for one specific
probability assignment. We had to roll four dice a bunch of times,
keeping track of the average of the dice for each roll, as well as the
total times that each number was rolled.

While not a challenging assignment, I found it really tedious, so
instead of rolling dice, I threw this program together and used it to
do most of the work for me instead. :P